buat file hapus.php

+++ latihan/index.php
@@ -29,7 +29,7 @@ $mahasiswa =query("SELECT * FROM mahasiswa");
     <tr>
       <td><?php echo $i; ?></td>
       <td><a href="ubah.php">Ubah</a> | 
-      <a href="hapus.php">Hapus</a>
+      <a href="hapus.php?id=<?= $row["id"];?>" onclick="return confirm('Yakin Hapus Data?');">Hapus</a>
       </td>
       <td><img src="images/<?=$row["gambar"]?>" width="50"></td>
       <td><?= $row["nim"]; ?></td>
